requirements.txt
jupyter-book
matplotlib
numpy
jupyterlite-sphinx
jupyterlite-pyodide-kernel